backend/requirements.txt
func_timeout==4.3.5
loguru
tqdm
tenacity
pandas
fastapi
uvicorn
pytest
pydantic_settings
aiohttp
httpx
nest_asyncio
datasets
mypy
types-setuptools
types-tqdm
psutil
openai>=1.0.0